Ceti Homeworks

+++ CETI/Programación-2/Mate.h
@@ -25,7 +25,7 @@ public:
     void EvenOrOdd();
     int ContadorCifras(int n);
     void DesplegarDatos(int n_datos, int *max, int *min, int *sum);
-    int CalcularPromedio();
+    float CalcularPromedio();
 };
